\name{binpar}
\alias{binpar}
\title{Starting Values of Parameters for the Binomial Data Set}
\usage{data(binpar)}
\description{
Starting values of parameters for fitting a mixture distribution to the data set \code{bindat}.

The \code{binpar} data frame has 4 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{pi}{the starting values for proportions.}
    \item{mu}{the starting values for means.}
    \item{sigma}{the starting values for standard deviations.}
  }
}
\examples{
data(binpar)
}
\keyword{datasets}
