\name{fiftn80}
\alias{fiftn80}
\title{A Mixed Data with Fifteen Normal Components}
\usage{data(fiftn80)}
\description{
Fifteen normal components grouped over eighty intervals.

The \code{fiftn80} data frame has 80 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{the boundaries of grouping intervals.}
    \item{freq}{the frequencies of observation falling into each interval.}
  }
}
\details{
A total of 820 observations were generated by computer to follow the distribution 
1/15 N(5, 1) + 1/15 N(10, 1) + ... + 1/15 N(75, 1) where N(\emph{m}, \emph{s}) denotes a normal 
distribution with mean \emph{m} and standard deviation \emph{s}.
}
\source{
  \url{http://www.math.mcmaster.ca/peter/mix/demex/ex1580.html}
}
\examples{
data(fiftn80)
plot.mixdata(fiftn80)
}
\keyword{datasets}
