\name{normals}
\alias{normals}
\title{Scale Mixture Data with Three Normal Components}
\usage{data(normals)}
\description{
Scale mixture of three normal distributions.

The \code{normals} data frame has 25 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{the boundaries of grouping intervals.}
    \item{freq}{the frequencies of observation falling into each interval.}
  }
}
\details{
A total of 249 observations were generated by computer to follow the mixture distribution 
1/3 N(12.5, 1) + 1/3 N(12.5, 3) + 1/3 N(12.5, 5) where N(\emph{m}, \emph{s}) denotes a 
normal distribution with mean \emph{m} and standard deviation \emph{s}.
}
\source{
  \url{http://www.math.mcmaster.ca/peter/mix/demex/exscle.html}
}
\examples{
data(normals)
plot.mixdata(normals)
}
\keyword{datasets}
