\name{pikeraw}
\alias{pikeraw}
\title{A Sample of Pike Lengths}
\usage{data(pikeraw)}
\description{
The data give the lengths of 523 pike (Esox lucius), sampled 
in 1965 from Heming Lake, Manitoba, Canada. There are known 
to be five age-groups in the sample. For each fish, the age 
group is determined by scale reading.

The \code{pikeraw} data frame has 523 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{length}{the lengths of 523 pike}
    \item{age}{the age groups of 523 pike}
  }
}
\source{
    Macdonald, P.D.M. and T.J. Pitcher (1979). 
    \emph{Age-groups from size-frequency data: a versatile and efficient method of analysing distribution mixtures.}
    Journal of the Fisheries Research Board of Canada 36, 987-1001.
}
\references{
    Macdonald, P.D.M. (1987). \emph{Analysis of length-frequency distributions.} 
    In R.C. Summerfelt and G.E. Hall [editors], Age and Growth of Fish, Iowa State University Press, Ames, Iowa. pp 371-384.

    \url{http://www.math.mcmaster.ca/peter/mix/demex/expike.html}
}
\examples{
data(pikeraw)
}
\keyword{datasets}
