\name{poispar}
\alias{poispar}
\title{Starting Values of Parameters for the Poisson Data Set}
\usage{data(poispar)}
\description{
Starting values of parameters for fitting a mixture distribution to the data set \code{poisdat}.

The \code{poispar} data frame has 4 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{pi}{the starting values for proportions.}
    \item{mu}{the starting values for means.}
    \item{sigma}{the starting values for standard deviations.}
  }
}
\examples{
data(poispar)
}
\keyword{datasets}
